services/vision-service/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
anthropic>=0.28.0
pillow>=10.0.0
pytesseract>=0.3.10
pyautogui>=0.9.54
